+++ tests/integration/fixture-theme/inc/qua-ham.php
@@ -1,0 +1,2 @@
+<?php
+const FXT_QUA_HAM = 1;